docs: updated scripts/data for manuscript
AMiGA is introduced and described by a manuscript to be posted on bioRxiv shortly (Midani, Collins, Britton. AMiGA: software for automated analysis of micorobial growth assays). All scripts/data needed to repeat the analysis in the manuscript and generate the figures are now included in the repository. Instructions are included in ./examples/manuscript/README.md