fix: zero or negative values are now handled differently
Previously, AMiGA would handle zero and negative values (at any time points) by simply raising them to a pre-defined threshold in config.py. However, this resulted in an issue where k_lin and auc_lin estimates (carrying capacity and area under the curve in linear OD units) would be undefined and thus zero, if the baseline is zero. This occuered because AMiGA was handling the zeros with raise_non_pos() but not updating the baseline attribute. So log-transformation followed by baseline subtraction did not prompt any errors, and growth curves can be fit just fine. However, in linearizing GP_Output (predictions of OD but in log-space) to linear OD, AMiGA was trying the following equation f(x) = exp(OD + ln(baseline)), which resulted in undefined valus if baseline is zero. This commit describes a fix where now AMiGA handldes zeros and negatives a bit differnetly (raise_non_pos() to handle_non_pos) and also includes an OD_Offset variable in keys which is instead used as the baseline for linearizations.